refactor(models): mirror backend model-registry SSOT on the frontend
Introduce src/lib/models/modelRegistry.ts as the single source of truth
for the 9 canonical segmentation models on the frontend, mirroring the
backend SSOT (backend/src/constants/modelRegistry.ts). Each registry
entry carries the full FE display metadata (size, defaultThreshold,
category, performance, English name/displayName/description, i18nKey) and
compatibleProjectTypes. ModelType, ALL_MODEL_IDS, BASE_MODEL_INFO,
BASIC_MODEL_INFO, the i18n keyMap, and the inverted
MODEL_TYPE_COMPATIBILITY are all DERIVED from it so they can never drift.

Re-point src/lib/modelUtils.ts to build getLocalizedModelInfo() /
getAllLocalizedModels() and re-export BASIC_MODEL_INFO / ModelType /
ModelInfo / ModelCategory / ModelPerformance from the registry. All
public export signatures are unchanged; the view-only SPHEROID_PRESETS
framing stays here (it is not model identity). Consumers are untouched.

Re-point src/types/index.ts KnownModelId + MODEL_TYPE_COMPATIBILITY to
derive from the FE registry; PROJECT_TYPES / ProjectType / isProjectType
are left exactly as-is.

Since both FE and BE now DERIVE MODEL_TYPE_COMPATIBILITY from their
registries (no literal object block on either side), drop the obsolete
MODEL_TYPE_COMPATIBILITY entry from scripts/verify-shared-types.cjs — its
regex can no longer locate a const block to diff. PROJECT_TYPES is still
cross-checked.

Add src/lib/models/__tests__/modelRegistry.test.ts asserting the
canonical 9, declaration order, the verified compatibility matrix, the
full ModelInfo shape, and exact per-model metadata. Pure refactor: zero
behavior change (existing modelUtils + types runtime tests still green).
